Rename Agda.Utils.ProfileOptions to Agda.Interaction.Options.ProfileOptions
This is an Agda-specific module rather than a general tool,
so it should not reside in the Utils/ tree where we keep general tools.
