+++ src/full/Agda/TypeChecking/Monad/Benchmark.hs
@@ -24,7 +24,7 @@ import qualified Agda.Utils.Benchmark as B
 
 import Agda.Utils.Monad
 import Agda.Syntax.Common.Pretty (prettyShow)
-import qualified Agda.Utils.ProfileOptions as Profile
+import qualified Agda.Interaction.Options.ProfileOptions as Profile
 
 -- | When profile options are set or changed, we need to turn benchmarking on or off.
 updateBenchmarkingStatus :: TCM ()
