+++ src/full/Agda/TypeChecking/Serialise/Instances/Errors.hs
@@ -19,7 +19,7 @@ import Agda.Interaction.Options.Warnings
 import Agda.Interaction.Library.Base
 import Agda.Termination.CutOff
 import Agda.Syntax.Common.Pretty
-import Agda.Utils.ProfileOptions
+import Agda.Interaction.Options.ProfileOptions
 
 import Agda.Utils.Impossible
